pyproject.toml
[tool.poetry]
name = "s3-tool"
version = "0.2.0"
description = "A comfortable CLI tool for working with AWS S3 buckets"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.0"
click = "^8.1.7"
python-dotenv = "^1.0.0"
python-magic = "^0.4.27"
requests = "^2.31.0"
boto3-stubs = "~=1.42.79"
botocore = "^1.42.79"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[tool.poetry.scripts]
s3-tool = "s3_tool.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"